Refactor: extract launch_aicpu + shared collectors/flags into base
Move the trivially-shared Group E pieces from a2a3 and a5 onboard
DeviceRunners into the shared DeviceRunnerBase (-10 net lines on what
already existed, with much of the structure now centralized):

- `launch_aicpu_kernel`: identical 3-line wrapper around
  `load_aicpu_op_.LaunchBuiltInOp`.
- Shared collector fields: `l2_perf_collector_`, `dump_collector_`,
  `pmu_collector_`, `scope_stats_collector_`.
- Shared enable flags + resolved values: `enable_l2_swimlane_`,
  `enable_dump_tensor_`, `enable_pmu_`, `enable_scope_stats_`,
  `l2_perf_level_`, `pmu_event_type_`, `output_prefix_`.
- Shared setters/getter: `set_l2_swimlane_enabled`,
  `set_dump_tensor_enabled`, `set_pmu_enabled`,
  `set_scope_stats_enabled`, `set_output_prefix`, `output_prefix()`.

Leaves arch-specific:
- a2a3: `enable_dep_gen_`, `dep_gen_collector_`, `set_dep_gen_enabled`,
  `init_dep_gen` (a2a3 only).
- Each arch's `init_l2_perf` / `init_tensor_dump` / `init_pmu` /
  `init_scope_stats` / `finalize_collectors` (divergent callback styles
  — a2a3 wraps halHostRegister/Unregister; a5 uses direct rtMalloc/rtFree).

Verbatim move. Both arches built clean. a2a3 onboard smoke 8/8 in 16s.
