+++ src/common/platform/onboard/host/device_runner_base.cpp
@@ -563,3 +563,17 @@ BindCallableResult DeviceRunnerBase::bind_callable_to_runtime(Runtime &runtime, 
         static_cast<int>(state.signature.size())
     };
 }
+
+// =============================================================================
+// Group E (minimal) — shared AICPU launch helper
+// =============================================================================
+
+int DeviceRunnerBase::launch_aicpu_kernel(
+    rtStream_t stream, KernelArgs *k_args, const char *kernel_name, int aicpu_num
+) {
+    // kernel_name is host::KernelNames::InitName / RunName — the runtime SO's
+    // actual exported symbol (simpler_aicpu_init / simpler_aicpu_exec).
+    // LaunchBuiltInOp dispatches via rtsLaunchCpuKernel on the cached
+    // rtFuncHandle resolved by LoadAicpuOp::Init at first-time bootstrap.
+    return load_aicpu_op_.LaunchBuiltInOp(stream, k_args, aicpu_num, kernel_name);
+}
